Add files via upload
This is an EAC-CPF 1.0 file for transformation testing. It includes all elements with all attributes, repeating those that can be repeated at least once. For the values of attributes and the content of elements, the name of the attribute or element has been used adding numbers in case of repetition, e.g. `<p  xml:id="source2-descriptiveNote-p1_ID" xml:lang="spa">source2-descriptiveNote-p1</p>` for the first paragraph within a descriptive note along side the second source mentioned in the `<control>` section. This is to more easily identify the origins of values and content in cases where EAC-CPF 2.0 will use these in other contexts or with renamed elements.

Status on upload (13 January 2021, 12.30pm, CET): The file is complete for the `<control>` section and includes the general structure for `<multipleIdentities>` with two `<cpfDescription>`-s.