Update EAC-CPF1_0.xml
Added `<relations>` section with nine variations for `<cpfRelation>` covering all options for `@cpfRelationType` and three variations each for `<resourceRelation>` and `<functionRelation>` covering all options for `@resourceRelationType` and `@functionRelationType`. This also includes variations of the date elements and a mix of `<objectBinWrap>` and `<objectXMLWrap>`.

Only missing now: `<alternativeSet>` in the second `<cpfDescription>`.